vtk5: remove formula
VTK 5.10.0 was released on Oct 17, 2012
VTK 6.0.0 was released on Jun 21, 2013

The vtk5 formula was moved from homebrew-versions to
homebrew-science, see commit a98415fa71f3ed2375fe60be27bca2d93aa0f9ea
There is no explanation in the commit message for this; but
the guess is that as there were big API changes between
vtk5 and vtk6, and it was probably good to keep
a vtk5 formula here.

Since then, vtk7 has been released. Most people seem to
have done the transition. Looking at the download statistics
on bintray for the vtk5 bottles, the number of downloads
has been very low for the last year.

If somebody really needs vtk5, they can still maintain this
formula in a separate tap.