vtk: improve libPythonX.Y.so detection
Under Ubunutu Xenial (but probably other distributions),
when no brewed python is installed, the actual logic
was picking up the wrong path for the libPythonX.Y.so
file.

The file is located under /usr/lib/x86_64-linux-gnu.

When no file is found, an error is now thrown,
to bail out early with a clear error message,
instead of having to read cryptic error messages
in the build log the next time this happens.

To my knowledge there seem to be no way to get this
path programatically. sysconfig.get_config_var() could
be used (checking for LIBDIR, LIBPC), but these values
seem inconsistent on OS X, so it is probably better
to handle this by trial and error.

Also, the caveats message was removed, as this is not true
anymore. We do not link against a specific python.
This should have been removed in commit
fdbb385236a86153444847eecb5831cc433eafe0
but was forgotten.